refactor: 패키지 구분하기 domain과 view로
