+++ src/main/java/domain/Calculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package domain;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
