+++ src/main/java/domain/Cars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.ArrayList;
 import java.util.Comparator;
