+++ src/test/java/calculator/CalculatorTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import domain.Calculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
