+++ src/test/java/racingcar/CarTest.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
