+++ src/test/java/racingcar/InputViewTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import view.InputView;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
