+++ src/test/java/racingcar/RacingCarTest.java
@@ -1,11 +1,9 @@
 package racingcar;
 
+import domain.Car;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.Random;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
