+++ src/test/java/racingcar/ResultViewTest.java
@@ -1,7 +1,10 @@
 package racingcar;
 
+import domain.Car;
+import domain.Cars;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import view.ResultView;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
